Implemented Average Normalizer
