+++ probs/relangProbs.py
@@ -1,65 +1,9 @@
 import json
-from collections import Counter
+import sys
 
-# Calculates average occurrences of types of a given feature in a sound inventory
-def averager(phonology: list[int], feature: int, averages: dict[int, int]) -> dict[int, float]:
-    feature_cats = Counter()
-    
-    for phoneme in phonology:
-        if feature != (7 << 13) or phoneme & (1 << 13) != 0:
-            feature_cats[phoneme & feature] += 1
-        
-        else:
-            feature_cats[0] += 1
-            
-
-    for cat in feature_cats:
-        if cat in averages:
-            averages[cat] += feature_cats[cat] / feature_cats.total()
-        
-        else:
-            averages[cat] = feature_cats[cat] / feature_cats.total()
-
-    return averages
-
-
-# Takes in the base probability and average found for a given category
-# of a feature and outputs a modded prob. 
-def probModder(base_probs: list[list[int, float]], avgs: dict[int, float], len_phono: int) -> list[list[int, float]]:
-    unmodded = []
-    total_diff = 0
-
-    # Modify probabilities of feature categories found in the given phonologies
-    for base_prob in base_probs:
-        if base_prob[0] in avgs:
-            
-            avg = avgs[base_prob[0]] / len_phono
-            mod = base_prob[1] - avg
-
-            base_prob[1] += mod
-            total_diff += mod
-
-        else:
-            unmodded += [base_prob[0]]
-
-    # If modding the averages caused them to not add up to 1, distribute the discrepancy
-    # across unmodded probs if possible, but all the probs otherwise.
-    if abs(total_diff) > 0.0001:
-        if unmodded:
-            
-            mod = total_diff * -1 / len(unmodded)
-            for base_prob in base_probs:
-
-                if base_prob[0] in unmodded:
-                    base_prob[1] += mod
-
-        else:
-            mod = total_diff * -1 / len(base_probs)
-            for base_prob in base_probs:
-                base_prob[1] += mod
-
-
-    return base_probs
+sys.path.append("..")
+from probs.calcAverage import averager, normalizer
+from probs.modProbs import probModder
 
 """
 This part of the Relang pipeline takes in all of the user-inputted phonologies and finds the
@@ -68,7 +12,7 @@ manners, laryngeal features). These averages ideally shine light into the user's
 constructing proto-phonologies. Subsequently, the function uses these averages to modify the
 base probs to steer away from those tendencies. 
 """
-def relangProbs(phonologies: list[list[int]]) -> dict[str, dict[str, list[list[int, float]]]]:
+def relangProbs(phonologies: list[list[int]]) -> dict[str, list[list[int, float]]]:
 
     # Calculate average percentages that places, manners, laryngeal features, etc. occur.
     # [place, laryngeals, manners, nasality, laterality, suprasegmentals]
@@ -86,6 +30,7 @@ def relangProbs(phonologies: list[list[int]]) -> dict[str, dict[str, list[list[i
 
     features = ["Place", "Laryngeals", "Manner", "Nasality", "Laterality", "Suprasegmentals"]
     for i in range(len(features)):
-        probModder(probs[features[i]], averages[i], len(phonologies))
+        averages[i] = normalizer(averages[i], probs[features[i]])
+        probs[features[i]] = probModder(probs[features[i]], averages[i], len(phonologies))
 
     return probs
