+++ src/main.py
@@ -64,8 +64,8 @@ if __name__ == "__main__":
 
         sel_phones = []
         if use_relang:
-            p = ["m", "n", "ŋ", "p", "t", "k", "s", "w", "l", "j"]
-            q = ["b", "d", "g", "p", "t", "k", "s", "w", "l", "j"]
+            p = ["m", "n", "ŋ", "p", "t", "k", "s", "ʋ", "l", "j"]
+            q = ["b", "d", "g", "p", "t", "k", "s", "ʋ", "l", "j"]
             phonos = [p, q]
 
             sel_phones = main(phonos)
